Fix some warnings
1. Removed unused variable in vpd_tool_impl.cpp.
2. Fix for sign compare warning in editor_impl.cpp & ibm_vpd_app.cpp.
3. Fix for unused variable warning in vpdecc files.

+++ ibm_vpd_app.cpp
@@ -260,7 +260,7 @@ Binary getVpdDataInVector(nlohmann::json& js, const string& filePath)
     vector<unsigned char>::const_iterator first = tmpVector.begin();
     vector<unsigned char>::const_iterator last = tmpVector.begin() + NEXT_64_KB;
 
-    if (distance(first, last) < tmpVector.size())
+    if (distance(first, last) < static_cast<int>(tmpVector.size()))
     {
         Binary vpdVector(first, last);
         return vpdVector;
